ci: publish:versions: use privileged SSH key instead of token
The token is not privileged as is meant to be used on status APIs. Use
the SSH key that we have across the Mender group to avoid introducing a
different privileged token and a risk of leak by mistake when rotating
them.

By the looks of it, we have never run this step fully automated so this
went unnoticed until the time of tagging the ongoing release.

Other minor beautifications.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>