fix: Do not double count integration in changelogs and statistics.
When calling `release-tool.py -l --only-backend`, integration was not
listed, but this was fixed in 170f2db294ed (3.2.0). So in all versions
after that, we should remove the extra addition to avoid counting it,
and rendering it, twice.

This breaks rendering of 3.1.0 and older, but we don't need those
anymore, and in any case, it's best to use the tool in the relevant
release branch when rendering.

Changelog: Title
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>