ci: Use GitHub token in addition to SSH key and fix hub command
This commit partially reverts 7bb7fc0. It uses a different token because
the previous one only had repo:status permissions.

It took me a while to realize that we actually need both the token and
the SSH key. See this comment:
* https://github.com/github/hub/issues/1644#issuecomment-359002547

Additionally, fix the hub command: a pull request needs to be launched
from a different fork or branch.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>