test: Do not reuse CompatibilySetup for Mender 2.5 setup
There are a set of tests that specifically verify `mender-connect` 1.0
with current server versions, and for historical reasons were reusing
the `DockerComposeCompatibilitySetup`. Leave the latter for the rather
specific test setup used there, and create new (smipler?) setups for the
tests requiring only this specific client.

This commit workarounds the problem for QA-525 by not using this setup
in other tests. The root cause for QA-525 is still unknown.

Changelog: None
Ticket: QA-525

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>