test: Do not use global devauth in any test in test_mender_connect
Depending on the xdist ordering for the function and class scoped tests,
the state of the global devauth that is carried along (no
`reset_mender_api` call in between) will make consecutive tests fail.

Fix by attaching new instances of `auth` and `devauth` to the OS test
environments (same way we are already doing for Enterprise tests).

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>