test: Improve robustness of test_keepalive_disable
The mender client will always do an inventory update on boot, which is
what presumably that 1s sleep was trying to wait for.

Increase considerably the sleep, while making the poll intervals long
so that no new inventory update (nor update check) happen for the rest
of the test.

This does not modify the goal of the test, which is to verify that the
client respects keep alive settings (specifically, that disabling it
does not keep TCP connections lingering around).

Ticket: QA-461

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>