test: Fix and enable `TestClientCompatibility`
It is unknown at which point in time started to fail - the assumption is
that an update in upstream `docker-compose` made this error reveal
itself.

Basically the test was combining `up ... scale=0` with `run ...` which
in effect would create two `mender-client` services. Although one will
still be not running, the later logic to check that we have only one
client would fail.

Moved also the TENANT_TOKEN to a shell variable because `up` command,
unlike `run`, does not support `-e` options in the command line.

Ticket: QA-525

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>