test: Top timeouts to 10 minutes
The one hours timeouts are ancient, they come from the Jenkins time.
With such long timeouts, it is impossible to run the full suite when
there are bugs in the software (the CI jobs will timeout after 8h and no
results will be gathered).

My reasoning is that if a devices takes more than 10 minutes to respond
to something, we have bugs in the software or we have such an overloaded
CI that we might as well just abort.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>