chore: Enable the mtls integration test
This enables the test, and makes it ready for the new C++ client.

This now tests the client with the OpenSSLv3 providers, as opposed to the old
test, which was testing the Engine API, which is now deprecated in OpenSSLv3,
and we are not compiling with deprecation support atm.

There are also some cosmetic changes. The previous setup used the OpenSSL config
to enable the Engine support.

This I have removed. We are not using the Engine API for this test in any case.

Our OpenSSLv3 Providers are loaded in code explicitly (which means they have to
be present in /usr/lib/ossl-modules), and not implicitly loaded through the
OpenSSL config - although this is still possible of course, if you don't place
it in the default lookup path mentioned previously.

It also changes the custom config given to `SoftHSM` - which is the emulator
used for the hardware crypto (HSM).

We now rely on the standard config, while previously we stored it in a custom
folder, which we deleted after the test, now we just delete them from the
standard location.

Now - let's pray to the CI gods.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>