feat(license-overview-generator): Handle repos migrated to `go-licenses`
These repos do not have anymore the checksum license file and instead
rely on `go-licenses` tool to generate the license information.

Use a custom template to list the components and licenses, and parse it
back to our custom dictionary, so that it can be collected for all
repositories nicely.

Changelog: None
Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>