test: Re-enable (and modify) `test_in_poor_network_environment`
The original test design assumed that `mender-connect` could reuse the
same TCP connection after the network is restored. However, the
`http-forwarder` in the Mender client (`mender` or `mender-auth`) do
close the TCP connection when we mess it up with `iptables`.

Modify the test to expect `mender-connect` to be able to open a new
shell after the connection is restored, which is evidence that it can
"auto-heal" after a network disruption.

`mender-connect` 2.2.0 and older do not pass this test due to MEN-6888.
It is fixed now at:
* https://github.com/mendersoftware/mender-connect/pull/130

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 135fc88f7d5e0a99079607f0d763200ea8c1da7c)