test: Fix test and honestly workaround the buggy behaviour
Previously, the test was filtering for `mender`, which in the span of
few minutes can trigger hundreds of alerts. The test was expecting more
than one alert and concluded that it worked fine.

This was not only flooding the mail server and consuming all logging
space in CI, but most importantly was hiding the fact that alerts can
often go missing when using this method.

Ticket: QA-543

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>