fix(licenses): Respect directories precedence
According to the tool's help text, the directories passed with `--dir`
should get precedence over the hard-coded relative path to the tool.

This was not the case, so the licenses generation was always called from
the local checkout (when the repo existed) over the correct versioned
temporary checkout prepared by the release tool.

Changelog: None
Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>