test: Make large image larger
For scarthgap we have increased the rootfs size to around 400M. This
test image then needs to be even larger to guarantee that it does not
fit.

Cherry-picking to 3.7.x is moving the fix to a different file (wrapper
instead of pytest fixture).

Ticket: MEN-7302

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 077b28469aee65cbf0636128260ab192538ebacc)