chore: Remove `mender-gateway` Docker image from release_tool
We introduced the image in 408078aa and has not yet been released. We
need to remove it because the build and publish of it is happening
already on its own repository, and involving the release_tool for it
will conflict (namely: it will override the multiplatform image with a
x86 only image).

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>