ci: Revert to hard-coded repository for mender-gateway QEMU image
This image, together with all the client ones, is still built from
mender-qa and parsed by the release_tool for the actual release.

As we want to limit the changes on the tool (and rather deprecate it
soon) we need to revert to the hard-coded Docker repository so that it
can be correctly parsed during release.

Moving also the existing client variable in the ci and env files to
group the images that are built from mender-qa pipeline.

Ticket: QA-794

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>