test: Improve test to catch bugs with the state machine hanging
The test was way to naive to test that we have "less than X loops". This
statement cannot catch bugs where the machine does not loop (iow does 1
loop).

Set the min to 4 to ensure at least two loops for each of the events:
polling for update and sending inventory.

This test fix would have caught bug MEN-7900.

Ticket: MEN-7900

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>