test: Update image used in tests using Docker client
To use the image that we will soon promote to "Virtual Device" for our
Get Started guides in the docs.

For the purpose of the tests, it should not really matter if the virtual
device is client only or also has the add-ons. By repurposing these
tests to use the image with add-ons we indirectly check that this image
is healthy.

Ticket: MEN-7634

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit f747c663d1da44a11dc58dfb34ba6e597db6000d)