test: refactor fs extraction and fix race condition
* Simplify  extraction by extracting it directly to the destination path
  and attempt to improve the error handling
* Move away from the request-based cleanup and do a simple yield
  instead

Previously the lock was only taken to check if the path existed, meaning
that if it didn't exist the lock would be released and multiple
workers could potentially enter `_image` simultaneously, resulting in
possible undefined behaviour

Remove the `run_if_non_existent` and simply call `image`. The previous
implementation created a lock in a tmp directory shared by the workers,
but we can just create a lock in the working directory and not have to
bother with worker ids etc.

Ticket: QA-1107

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>