test: retry DnsRequest in test_portforward
Wrap `DnsRequest` with `redo.retry` in order to ensure that the test
doesn't fail just because the portforwarding isn't ready.

I was able to reproduce the SocketError by removing the previous sleep;
hopefully this is what's causing the SocketErrors in CI, and adding
retries and sleeps in between should hopefully mitigate the test
failures.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>