fix: fine-tune docker images priorities to fix failing test:integration:1 pipeline
A lot of tests manipulate artifacts and make rootfs updates, which is very heavy on disk I/O. This caused
tenantadm to time out on creating organization sometimes. Changed docker containers I/O priorities and moved mongodb
database directory to RAM. This made some operations much faster, and the error almost dissapeared.
Unfortunately, not 100% of issues were possible to fix. For example, sometimes mender-workflows-server gets a
"no route to host" error when dispatching a job to workflows-worker. This is unrelated to the disk I/O issue
but results in tenantadm failing in the same way.

Ticket: QA-1108
Changelog: Title

Signed-off-by: Michal Kopczan <michal.kopczan@northern.tech>