ci: exit with error message if branch is not protected
Instead of getting vague messages like
`Error: Cannot perform an interactive login from a non TTY device`
because the secrets aren't available, we'll exit early and print a
message explaining why.

The integration tests need to be run in a protected branch. We could
add a rule and not run the job at all, but that could cause a pipeline
with `RUN_TESTS_FULL_INTEGRATION=true` to be green even if the tests
aren't run

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>