test: update tests for updated abortion handling
Allow specifying the number of reboots expected when aborting a
deployment. A reboot after install has been posted will only reboot once
as it goes straight to rollback reboot. Remove mender_performs_reboot
and use number_of_reboots to determine whether or not to it should
expect a reboot.

Don't expect a reboot of the client if the deployment is aborted when
sending "downloading"

Set the polling_frequency in check_expected_statistics to 0 to make sure
we catch the abort steps for `installing`.

Ticket: ME-527

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>