test: use requests.get rather than curl
We've started seeing failures like:
`OSError: [Errno 8] Exec format error: '/tmp/tmphyozt1q2/gen_docker-compose'`
This is happening in the newly introduced docker-compose tests.
To make sure we're not failing because of curl and to do it in a more pythonic way,
we now use `requests.get` decorated with a retrier to try to ensure we actually
have the script properly available before trying to execute it.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>